adjustments to make ravendb core tests pass on Linx & mono

+++ Raven.Database/Storage/Voron/TransactionalStorage.cs
@@ -66,7 +66,7 @@ namespace Raven.Storage.Voron
 
 			documentCacher = new DocumentCacher(configuration);
 			exitLockDisposable = new DisposableAction(() => Monitor.Exit(this));
-            bufferPool = new BufferPool(configuration.Storage.Voron.MaxBufferPoolSize * 1024 * 1024 * 1024, int.MaxValue); // 2GB max buffer size (voron limit)
+			bufferPool = new BufferPool((long)configuration.Storage.Voron.MaxBufferPoolSize * 1024 * 1024 * 1024, int.MaxValue); // 2GB max buffer size (voron limit)
 		}
 
 		public void Dispose()
